feat(RECIPE-21) add JWT authentication and authorization for recipe management

+++ src/main/java/dika/recipeservice/service/impl/RecipeServiceImpl.java
@@ -1,7 +1,6 @@
 package dika.recipeservice.service.impl;
 
 
-import dika.recipeservice.rabbit.RabbitEventPublisher;
 import dika.recipeservice.dto.RecipeCreateDto;
 import dika.recipeservice.dto.RecipeDto;
 import dika.recipeservice.dto.RecipePageDto;
@@ -10,14 +9,20 @@ import dika.recipeservice.enums.Status;
 import dika.recipeservice.exception.RecipeNotFound;
 import dika.recipeservice.mapper.RecipeMapper;
 import dika.recipeservice.model.Recipe;
+import dika.recipeservice.rabbit.RabbitEventPublisher;
 import dika.recipeservice.repository.RecipeRepository;
+import dika.recipeservice.service.JwtService;
 import dika.recipeservice.service.RecipeService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
+import java.util.UUID;
 import java.util.function.Consumer;
 
 
@@ -28,6 +33,7 @@ public class RecipeServiceImpl implements RecipeService {
     private final RecipeRepository recipeRepository;
     private final RecipeMapper recipeMapper;
     private final RabbitEventPublisher rabbitEventPublisher;
+    private final JwtService jwtService;
 
     @Transactional(readOnly = true)
     public RecipePageDto getAllRecipes(Pageable pageable) {
@@ -37,13 +43,21 @@ public class RecipeServiceImpl implements RecipeService {
 
     @Transactional
     public RecipeDto createRecipe(RecipeCreateDto recipeDto) {
+        Recipe recipe = recipeMapper.toEntityRecipeCreateDto(recipeDto);
+        recipe.setAuthorExternalId(getExternalId());
+        recipe.setAuthorUsername(getUsername());
         rabbitEventPublisher.publishRecipeCreatedEvent(recipeDto);
-        return recipeMapper.toDto(recipeRepository.save(recipeMapper.toEntityRecipeCreateDto(recipeDto)));
+        recipeRepository.save(recipe);
+        return recipeMapper.toDto(recipe);
     }
 
     @Transactional
     public RecipeDto update(Long id, RecipeCreateDto recipeDto) {
         Recipe recipe = recipeRepository.findById(id).orElseThrow(() -> new RecipeNotFound("Recipe not found"));
+        if (!recipe.getAuthorExternalId().equals(getExternalId())
+                || !recipe.getAuthorUsername().equals(getUsername())) {
+            throw new RecipeNotFound("You are not the author of this recipe");
+        }
         updateFields(recipeDto.description(), recipe::setDescription);
         updateFields(recipeDto.ingredients(), recipe::setIngredients);
         updateFields(recipeDto.title(), recipe::setTitle);
@@ -59,6 +73,11 @@ public class RecipeServiceImpl implements RecipeService {
 
     @Transactional
     public void deleteRecipe(Long id) {
+        Recipe recipe = recipeRepository.findById(id).orElseThrow(() -> new RecipeNotFound("Recipe not found"));
+        if (!recipe.getAuthorExternalId().equals(getExternalId())
+                || !recipe.getAuthorUsername().equals(getUsername())) {
+            throw new RecipeNotFound("You are not the author of this recipe");
+        }
         recipeRepository.deleteById(id);
     }
 
@@ -90,5 +109,19 @@ public class RecipeServiceImpl implements RecipeService {
         if (newParam != null) {
             oldParam.accept(newParam);
         }
+    }
+
+    private String getToken() {
+        HttpServletRequest request = ((ServletRequestAttributes) RequestContextHolder
+                .currentRequestAttributes()).getRequest();
+        return request.getHeader("Authorization").substring(7);
+    }
+
+    private String getUsername() {
+        return jwtService.extractUsername(getToken());
+    }
+
+    private UUID getExternalId() {
+        return jwtService.extractExternalId(getToken());
     }
 }
